+++ sdk/docs/ref/functions/typedCreateHandler.md
@@ -1,4 +1,4 @@
-[**@saflib/vue**](../../../../index.md)
+[**@saflib/sdk**](../index.md)
 
 ***
