+++ sdk/typedoc.json
@@ -1,0 +1,7 @@
+{
+  "intentionallyNotExported": [
+    "ExtractResponseBody",
+    "ExtractRequestBody",
+    "ExtractRequestParams"
+  ]
+}
