+++ vue/docs/ref/@saflib/vue/testing/index.md
@@ -20,5 +20,4 @@ Testing utilities for Vue.
 | [mountWithPlugins](functions/mountWithPlugins.md) | Mount a Vue component with plugins. Handles plugins like vuetify, router, and i18n. Uses `mount` under the hood. |
 | [setupMockServer](functions/setupMockServer.md) | Simple wrapper around `msw`'s `setupServer` function. |
 | [stubGlobals](functions/stubGlobals.md) | Call during test setup to stub browser globals like ResizeObserver, matchMedia, location, and visualViewport. |
-| [typedCreateHandler](functions/typedCreateHandler.md) | Use to create a typed helper function for creating typesafe mock API handlers. |
 | [withVueQuery](functions/withVueQuery.md) | Helper function to test Vue Query composables in isolation. |
